@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
-import { map, shareReplay, timeout } from 'rxjs/operators';
+import { map, shareReplay } from 'rxjs/operators';
 import { Coordinate } from '../model';
 
 @Injectable({
@@ -23,7 +23,6 @@ export class LocationService {
           lng: position.coords.longitude
         };
       }),
-      timeout(10000), // Sometimes we don't get a location. Which makes you think the UI is not working...
       shareReplay({ bufferSize: 1, refCount: true })
     );
   }
